ois window handle cannot be int,has to be long. this is bad syntax when checked.

+++ src/Core/Input.cpp
@@ -25,8 +25,7 @@ void Input::initOis(
 		OIS::KeyListener *pKeyListener,
 		OIS::MouseListener *pMouseListener
 	) {
-	// unsigned long hWnd = 0;
-	unsigned int hWnd = 0;
+	unsigned long hWnd = 0;
 
     RenderEngine::Instance().m_pRenderWnd->getCustomAttribute("WINDOW", &hWnd);
